+++ genbrush.h
@@ -625,7 +625,7 @@ void GBSurfaceSetFinalPixelSafe(GBSurface* const that,
 #if BUILDMODE != 0
 inline
 #endif 
-int GBSurfaceArea(const GBSurface* const that);
+long GBSurfaceArea(const GBSurface* const that);
 
 // Get the background color of the GBSurface 'that'
 #if BUILDMODE != 0
